requirements.txt
streamlit
chromadb==0.4.24
groq
python-dotenv
requests
beautifulsoup4
langchain-text-splitters
protobuf==4.25.3
onnxruntime
tokenizers
